Split context responsibilities into separate modules
