+++ modules/core/domRefs.js
@@ -1,0 +1,24 @@
+const domRefs = {
+  gameWrapper: null,
+  kakaoBox: null,
+  choiceContainer: null,
+  talkWrapper: null,
+  nameEl: null,
+  bubbleEl: null,
+  kakaoOverlay: null,
+  overlayImage: null,
+  skipBtn: null,
+  saveBtn: null,
+  answerUi: null,
+  answerInput: null,
+  submitAnswer: null,
+  hintPopup: null,
+  hintStep1: null,
+  hintStep2: null,
+  hintBtn: null,
+  hintConfirm: null,
+  msgBox: null,
+  savePopup: null,
+};
+
+export default domRefs;
